+++ service/orchestrator/orchestrator_test.go
@@ -275,112 +275,3 @@ func TestService_GetRuntimeInfo(t *testing.T) {
 		})
 	}
 }
-
-func TestService_ListPublicCertificates(t *testing.T) {
-	type fields struct {
-		UnimplementedOrchestratorServer orchestrator.UnimplementedOrchestratorServer
-		cloudServiceHooks               []orchestrator.CloudServiceHookFunc
-		toeHooks                        []orchestrator.TargetOfEvaluationHookFunc
-		AssessmentResultHooks           []func(result *assessment.AssessmentResult, err error)
-		storage                         persistence.Storage
-		metricsFile                     string
-		loadMetricsFunc                 func() ([]*assessment.Metric, error)
-		catalogsFolder                  string
-		loadCatalogsFunc                func() ([]*orchestrator.Catalog, error)
-		events                          chan *orchestrator.MetricChangeEvent
-		authz                           service.AuthorizationStrategy
-	}
-	type args struct {
-		in0 context.Context
-		req *orchestrator.ListPublicCertificatesRequest
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		wantRes *orchestrator.ListPublicCertificatesResponse
-		wantErr assert.ErrorAssertionFunc
-	}{
-		{
-			name:   "Validation error",
-			fields: fields{},
-			args: args{
-				req: nil,
-			},
-			wantRes: nil,
-			wantErr: func(tt assert.TestingT, err error, i ...interface{}) bool {
-				assert.Equal(t, codes.InvalidArgument, status.Code(err))
-				return assert.ErrorContains(t, err, api.ErrEmptyRequest.Error())
-			},
-		},
-		{
-			name: "Pagination error",
-			fields: fields{
-				storage: &testutil.StorageWithError{ListErr: ErrSomeError},
-			},
-			args: args{
-				req: &orchestrator.ListPublicCertificatesRequest{},
-			},
-			wantRes: nil,
-			wantErr: func(tt assert.TestingT, err error, i ...interface{}) bool {
-				assert.Equal(t, codes.Internal, status.Code(err))
-				return assert.ErrorContains(t, err, "database error")
-			},
-		},
-		{
-			name: "Happy path",
-			fields: fields{
-				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
-					// Create Certificate
-					assert.NoError(t, s.Create(orchestratortest.NewCertificate()))
-				}),
-			},
-			args: args{
-				req: &orchestrator.ListPublicCertificatesRequest{},
-			},
-			wantRes: &orchestrator.ListPublicCertificatesResponse{
-				Certificates: []*orchestrator.Certificate{
-					{
-						Id:             testdata.MockCertificateID,
-						Name:           testdata.MockCertificateName,
-						CloudServiceId: testdata.MockCloudServiceID1,
-						IssueDate:      "2021-11-06",
-						ExpirationDate: "2024-11-06",
-						Standard:       testdata.MockCertificateName,
-						AssuranceLevel: testdata.AssuranceLevelHigh,
-						Cab:            testdata.MockCertificateCab,
-						Description:    testdata.MockCertificateDescription,
-					},
-				},
-			},
-			wantErr: assert.NoError,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			svc := &Service{
-				UnimplementedOrchestratorServer: tt.fields.UnimplementedOrchestratorServer,
-				cloudServiceHooks:               tt.fields.cloudServiceHooks,
-				toeHooks:                        tt.fields.toeHooks,
-				AssessmentResultHooks:           tt.fields.AssessmentResultHooks,
-				storage:                         tt.fields.storage,
-				metricsFile:                     tt.fields.metricsFile,
-				loadMetricsFunc:                 tt.fields.loadMetricsFunc,
-				catalogsFolder:                  tt.fields.catalogsFolder,
-				loadCatalogsFunc:                tt.fields.loadCatalogsFunc,
-				events:                          tt.fields.events,
-				authz:                           tt.fields.authz,
-			}
-			gotRes, err := svc.ListPublicCertificates(tt.args.in0, tt.args.req)
-			assert.NoError(t, gotRes.Validate())
-
-			tt.wantErr(t, err)
-
-			if tt.wantRes != nil {
-				if !reflect.DeepEqual(gotRes, tt.wantRes) {
-					t.Errorf("Service.ListPublicCertificates() = %v, want %v", gotRes, tt.wantRes)
-				}
-			}
-		})
-	}
-}
